[TRAV-85] moving generate keys code to abstract super class
