+++ cheddar/cheddar-integration-aws/src/main/java/com/clicktravel/infrastructure/persistence/aws/dynamodb/AbstractDynamoDbTemplate.java
@@ -27,7 +27,9 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDbPropertyMarshaller;
 import com.amazonaws.services.dynamodbv2.model.*;
 import com.clicktravel.cheddar.infrastructure.persistence.database.AbstractDatabaseTemplate;
+import com.clicktravel.cheddar.infrastructure.persistence.database.GeneratedKeyHolder;
 import com.clicktravel.cheddar.infrastructure.persistence.database.Item;
+import com.clicktravel.cheddar.infrastructure.persistence.database.SequenceKeyGenerator;
 import com.clicktravel.cheddar.infrastructure.persistence.database.configuration.DatabaseSchemaHolder;
 import com.clicktravel.cheddar.infrastructure.persistence.database.configuration.ItemConfiguration;
 import com.clicktravel.cheddar.infrastructure.persistence.database.configuration.SequenceConfiguration;
@@ -42,6 +44,10 @@ public abstract class AbstractDynamoDbTemplate extends AbstractDatabaseTemplate 
     protected final DatabaseSchemaHolder databaseSchemaHolder;
     protected final HashMap<Class<? extends Item>, ItemConfiguration> itemConfigurationMap;
     protected final Set<String> sequenceConfigurations;
+
+    private static final String SEQUENCE_TABLE_NAME = "sequences";
+    private static final String SEQUENCE_NAME_ATTRIBUTE = "name";
+    private static final String SEQUENCE_CURRENT_VALUE_ATTRIBUTE = "currentValue";
 
     public AbstractDynamoDbTemplate(final DatabaseSchemaHolder databaseSchemaHolder) {
         this.databaseSchemaHolder = databaseSchemaHolder;
@@ -62,6 +68,37 @@ public abstract class AbstractDynamoDbTemplate extends AbstractDatabaseTemplate 
 
     public DatabaseSchemaHolder databaseSchemaHolder() {
         return databaseSchemaHolder;
+    }
+
+    @Override
+    public GeneratedKeyHolder generateKeys(final SequenceKeyGenerator sequenceKeyGenerator) {
+        final String sequenceName = sequenceKeyGenerator.sequenceName();
+        if (!sequenceConfigurations.contains(sequenceName)) {
+            throw new IllegalStateException("Unsupported sequence: " + sequenceName);
+        }
+        final Map<String, AttributeValue> key = new HashMap<>();
+        key.put(SEQUENCE_NAME_ATTRIBUTE, new AttributeValue(sequenceName));
+        final AttributeValueUpdate attributeValueUpdate = new AttributeValueUpdate().withAction("ADD").withValue(
+                new AttributeValue().withN(String.valueOf(sequenceKeyGenerator.keyCount())));
+        final Map<String, AttributeValueUpdate> attributeUpdates = new HashMap<>();
+        attributeUpdates.put(SEQUENCE_CURRENT_VALUE_ATTRIBUTE, attributeValueUpdate);
+        final String tableName = databaseSchemaHolder.schemaName() + "-" + SEQUENCE_TABLE_NAME;
+        final UpdateItemRequest updateItemRequest = new UpdateItemRequest().withTableName(tableName).withKey(key)
+                .withAttributeUpdates(attributeUpdates).withReturnValues("UPDATED_NEW");
+        final UpdateItemResult updateItemResult;
+        try {
+            updateItemResult = amazonDynamoDbClient.updateItem(updateItemRequest);
+        } catch (final AmazonServiceException e) {
+            throw new PersistenceResourceFailureException("Failure while attempting DynamoDb Update (generate keys)", e);
+        }
+        final Map<String, AttributeValue> attributes = updateItemResult.getAttributes();
+        final AttributeValue currentAttributeValue = attributes.get(SEQUENCE_CURRENT_VALUE_ATTRIBUTE);
+        final Long currentValue = Long.valueOf(currentAttributeValue.getN());
+        final Collection<Long> keys = new ArrayList<>();
+        for (long i = currentValue - sequenceKeyGenerator.keyCount(); i < currentValue; i++) {
+            keys.add(i + 1);
+        }
+        return new GeneratedKeyHolder(keys);
     }
 
     protected final ItemConfiguration getItemConfiguration(final Class<? extends Item> itemClass) {
