+++ cheddar/cheddar-integration-aws/src/main/java/com/clicktravel/infrastructure/persistence/aws/dynamodb/DynamoDbTemplate.java
@@ -28,7 +28,9 @@ import org.slf4j.LoggerFactory;
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDbPropertyMarshaller;
 import com.amazonaws.services.dynamodbv2.model.*;
-import com.clicktravel.cheddar.infrastructure.persistence.database.*;
+import com.clicktravel.cheddar.infrastructure.persistence.database.BatchDatabaseTemplate;
+import com.clicktravel.cheddar.infrastructure.persistence.database.Item;
+import com.clicktravel.cheddar.infrastructure.persistence.database.ItemId;
 import com.clicktravel.cheddar.infrastructure.persistence.database.configuration.*;
 import com.clicktravel.cheddar.infrastructure.persistence.database.exception.ItemConstraintViolationException;
 import com.clicktravel.cheddar.infrastructure.persistence.database.exception.NonExistentItemException;
@@ -42,9 +44,6 @@ public class DynamoDbTemplate extends AbstractDynamoDbTemplate implements BatchD
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     private static final String VERSION_ATTRIBUTE = "version";
-    private static final String SEQUENCE_TABLE_NAME = "sequences";
-    private static final String SEQUENCE_NAME_ATTRIBUTE = "name";
-    private static final String SEQUENCE_CURRENT_VALUE_ATTRIBUTE = "currentValue";
 
     public DynamoDbTemplate(final DatabaseSchemaHolder databaseSchemaHolder) {
         super(databaseSchemaHolder);
@@ -492,37 +491,6 @@ public class DynamoDbTemplate extends AbstractDynamoDbTemplate implements BatchD
         }
         final List<Map<String, AttributeValue>> itemAttributeMaps = batchGetItemResult.getResponses().get(tableName);
         return marshallIntoObjects(itemClass, itemAttributeMaps);
-    }
-
-    @Override
-    public GeneratedKeyHolder generateKeys(final SequenceKeyGenerator sequenceKeyGenerator) {
-        final String sequenceName = sequenceKeyGenerator.sequenceName();
-        if (!sequenceConfigurations.contains(sequenceName)) {
-            throw new IllegalStateException("Unsupported sequence: " + sequenceName);
-        }
-        final Map<String, AttributeValue> key = new HashMap<>();
-        key.put(SEQUENCE_NAME_ATTRIBUTE, new AttributeValue(sequenceName));
-        final AttributeValueUpdate attributeValueUpdate = new AttributeValueUpdate().withAction("ADD").withValue(
-                new AttributeValue().withN(String.valueOf(sequenceKeyGenerator.keyCount())));
-        final Map<String, AttributeValueUpdate> attributeUpdates = new HashMap<>();
-        attributeUpdates.put(SEQUENCE_CURRENT_VALUE_ATTRIBUTE, attributeValueUpdate);
-        final String tableName = databaseSchemaHolder.schemaName() + "-" + SEQUENCE_TABLE_NAME;
-        final UpdateItemRequest updateItemRequest = new UpdateItemRequest().withTableName(tableName).withKey(key)
-                .withAttributeUpdates(attributeUpdates).withReturnValues("UPDATED_NEW");
-        final UpdateItemResult updateItemResult;
-        try {
-            updateItemResult = amazonDynamoDbClient.updateItem(updateItemRequest);
-        } catch (final AmazonServiceException e) {
-            throw new PersistenceResourceFailureException("Failure while attempting DynamoDb Update (generate keys)", e);
-        }
-        final Map<String, AttributeValue> attributes = updateItemResult.getAttributes();
-        final AttributeValue currentAttributeValue = attributes.get(SEQUENCE_CURRENT_VALUE_ATTRIBUTE);
-        final Long currentValue = Long.valueOf(currentAttributeValue.getN());
-        final Collection<Long> keys = new ArrayList<>();
-        for (long i = currentValue - sequenceKeyGenerator.keyCount(); i < currentValue; i++) {
-            keys.add(i + 1);
-        }
-        return new GeneratedKeyHolder(keys);
     }
 
     /**
